requirements.txt
asgiref==3.8.1
click==8.1.7
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
django-import-export==4.2.1
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
Markdown==3.7
packaging==24.1
psycopg2-binary==2.9.10
sqlparse==0.5.1
tablib==3.7.0
typing_extensions==4.12.2
uvicorn==0.32.0
whitenoise==6.7.0
